# Beispielanwendung Adressbuch/Backend/src/api/openapi.yaml
openapi: 3.1.0
info:
    title: Adressbuch
    version: '1.0'
    description: >-
        Die Steuer-App nutzt einen simplen REST-Webservices zur Kommunikation mit dem Backend. Dieser
        ermöglicht die üblichen CRUD-Operationen zur Suche, Anzeige, und erstellung von Steuerjahren, wobei
        jedes Steuerjahr aus einem jahr, werbungskosten besteht.
    contact:
        name: Dominik Hoffmann
    license:
        name: CC-BY-4.0
        url: 'https://creativecommons.org/licenses/by/4.0/deed.de'
servers:
    -
        url: 'http://localhost:3000'
paths:
    /:
        get:
            tags:
                - Index
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/HATEOAS_Collection'
                                x-content-type: application/json
                    description: OK
            operationId: index
            summary: Wurzeladresse des Webservices
            description: Liefert eine HATEOAS-Übersicht aller collections
    /steuerjahr:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Steuerjahr'
            tags:
                - Steuerjahr
            responses:
                '201':
                    headers:
                        Location:
                            schema:
                                type: integer
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Steuerjahr'
                    description: Created
            operationId: createSteuerjahr
            summary: Steuerjahr anlegen
            description: >-
                Neues STeuerjahr anlegen. Speichert das übergebe Steuerjahr in der Datenbank. Der Wert des
                Feldes "_id" wird dabei ignoriert, da in jedem Fall eine neues Steuerjahr mit neuer ID
                angelegt wird.
    /steuerjahr/{id}:
        get:
            tags:
                - Steuerjahr
            parameters:
                -
                    style: simple
                    explode: false
                    name: id
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Steuerjahr'
                    description: OK
                '404':
                    description: Not Found
            operationId: getSteuerjahr
            summary: Steuerjahr lesen
            description: Liest eine einzelnes Steuerjahr vom Server.
components:
    schemas:
        HATEOAS_Collection:
            title: HATEOAS-Links einer Collection
            type: object
            properties:
                _name:
                    type: string
                query:
                    $ref: '#/components/schemas/HATEOAS_Link'
                create:
                    $ref: '#/components/schemas/HATEOAS_Link'
        HATEOAS_Resource:
            title: HATEOAS-Links einer Resource
            type: object
            properties:
                read:
                    $ref: '#/components/schemas/HATEOAS_Link'
                update:
                    $ref: '#/components/schemas/HATEOAS_Link'
                patch:
                    $ref: '#/components/schemas/HATEOAS_Link'
                delete:
                    $ref: '#/components/schemas/HATEOAS_Link'
        HATEOAS_Link:
            title: HATEOAS-Link
            type: object
            properties:
                url:
                    type: string
                method:
                    type: string
                query_params:
                    type: array
                    items:
                        type: string
        Steuerjahr:
            title: Steuerjahr
            type: object
            properties:
                _links:
                    $ref: '#/components/schemas/HATEOAS_Resource'
                _id:
                    type: string
                jahr:
                    type: integer
                werbungskosten:
                    type: integer
            example:
                _id: 621ff9a936997df5baba746e
                jahr: 2000
                werbungskosten: 15000
tags:
    -
        name: Index
    -
        name: Steuerjahr